Finished full test of finite element convergence for complex BVP
