+++ lib/lf/fe/test/CMakeLists.txt
@@ -21,3 +21,11 @@ target_link_libraries(lf.fe.fullgaltest PUBLIC
   lf.mesh.hybrid2d lf.mesh.hybrid2d lf.mesh.utils lf.mesh.test_utils lf.fe)
 target_compile_features(lf.fe.fullgaltest PUBLIC cxx_std_17)
 gtest_discover_tests(lf.fe.fullgaltest)
+
+add_executable(lf.fe.bvptest bvp_fe_test.cc)
+target_link_libraries(lf.fe.bvptest PUBLIC
+  Eigen3::Eigen Boost::boost GTest::main lf.mesh lf.refinement lf.assemble lf.quad
+  lf.mesh.hybrid2d lf.mesh.hybrid2d lf.mesh.utils lf.mesh.test_utils lf.fe)
+target_compile_features(lf.fe.bvptest PUBLIC cxx_std_17)
+gtest_discover_tests(lf.fe.bvptest)
+
