+++ lib/lf/fe/test/full_gal_test.cc
@@ -412,141 +412,4 @@ TEST(lf_full, bd_rhs_2) {
   EXPECT_NEAR(en.back(), 1.62539, 1.0E-4);
 }
 
-TEST(lf_gfe, Neu_BVP_ass) {
-  std::cout << "#### TEST: Assembly for Neumann BVP ###" << std::endl;
-  // Right-hand side source function
-  auto f = [](Eigen::Vector2d x) -> double {
-    return (std::sin(2 * M_PI * x[0]) * std::sin(2 * M_PI * x[1]));
-  };
-  // Neumann data
-  auto h = [](Eigen::Vector2d /*x*/) -> double { return 1.0; };
-
-  // Class describing the Neumann boundary value problem
-  std::shared_ptr<PureNeumannProblemLaplacian<decltype(f), decltype(h)>> bvp_p =
-      std::make_shared<PureNeumannProblemLaplacian<decltype(f), decltype(h)>>(
-          f, h);
-
-  // Obtain finite Element space
-  // Building the test mesh: a general hybrid mesh
-  auto mesh_p = lf::mesh::test_utils::GenerateHybrid2DTestMesh(0);
-
-  // Set up global FE space; lowest order Lagrangian finite elements
-  LinearLagrangianFESpace<double> fe_space(mesh_p);
-
-  // Generate linear system
-  auto [A, phi] = SecOrdEllBVPLagrFELinSys<double>(fe_space, bvp_p);
-
-  Eigen::MatrixXd A_dense = A;
-  std::cout << "Galerkin matrix = " << A_dense << std::endl;
-  std::cout << "rhs. vector = " << phi.transpose() << std::endl;
-  EXPECT_NEAR(A_dense.rowwise().sum().norm(), 0.0, 1.0E-10)
-      << "Row sum not zero";
-  EXPECT_NEAR(A_dense.colwise().sum().norm(), 0.0, 1.0E-10)
-      << "Col sum not zero";
-}
-
-// **********************************************************************
-// Test: Solution of BVP
-// **********************************************************************
-
-template <typename SOLFUNC, typename SOLGRAD>
-std::vector<std::pair<double, double>> TestConvergenceEllBVPFESol(
-    int reflevels, std::shared_ptr<const SecondOrderEllipticBVP<double>> bvp_p,
-    SOLFUNC solution, SOLGRAD sol_grad) {
-  std::vector<std::pair<double, double>> errnorms{};
-
-  // Generate hierarchy of meshes of the unit square
-  std::shared_ptr<lf::refinement::MeshHierarchy> multi_mesh_p =
-      lf::refinement::GenerateMeshHierarchyByUniformRefinemnt(
-          lf::mesh::test_utils::GenerateHybrid2DTestMesh(0, 1.0 / 3),
-          reflevels);
-  lf::refinement::MeshHierarchy& multi_mesh{*multi_mesh_p};
-  // Number of levels
-  size_type L = multi_mesh.NumLevels();
-
-  // Do computations on all levels
-  for (int l = 0; l < L; ++l) {
-    std::shared_ptr<const mesh::Mesh> mesh_p{multi_mesh.getMesh(l)};
-    // Set up global FE space; lowest order Lagrangian finite elements
-    LinearLagrangianFESpace<double> fe_space(mesh_p);
-    const lf::assemble::DofHandler& dofh{fe_space.LocGlobMap()};
-
-    // Compute Galerkin matrix A and right-hand-side vector
-    auto [A, phi] = SecOrdEllBVPLagrFELinSys<double>(fe_space, bvp_p);
-    // Solve linear system
-    Eigen::SparseLU<Eigen::SparseMatrix<double>> solver;
-    solver.compute(A);
-    Eigen::VectorXd sol_vec = solver.solve(phi);
-    // Compute error norms
-    // Helper class for L2 error computation
-    LocalL2NormDifference lc_L2(
-        fe_space.ShapeFunctionLayout(lf::base::RefEl::kTria()),
-        fe_space.ShapeFunctionLayout(lf::base::RefEl::kQuad()), solution);
-    // Helper class for H1 semi norm
-    LocL2GradientFEDifference lc_H1(
-        fe_space.ShapeFunctionLayout(lf::base::RefEl::kTria()),
-        fe_space.ShapeFunctionLayout(lf::base::RefEl::kQuad()), sol_grad);
-
-    double L2err = NormOfDifference(dofh, lc_L2, sol_vec);
-    double H1serr = NormOfDifference(dofh, lc_H1, sol_vec);
-    errnorms.emplace_back(L2err, H1serr);
-  }
-
-  return errnorms;
-}
-
-// Definition of boundary value problem
-
-class MixedEllipticBVP : public SecondOrderEllipticBVP<double> {
- public:
-  MixedEllipticBVP()
-      : edge_sel_(
-            [](Eigen::Vector2d x) -> bool { return ((x[0] + x[1]) < 1.0); }) {}
-  MixedEllipticBVP(const MixedEllipticBVP&) = default;
-  MixedEllipticBVP(MixedEllipticBVP&&) noexcept = default;
-  MixedEllipticBVP& operator=(const MixedEllipticBVP&) = default;
-  MixedEllipticBVP& operator=(MixedEllipticBVP&&) noexcept = default;
-
-  Eigen::Matrix<double, 2, 2> alpha(Eigen::Vector2d x) const override {
-    return (1 + x.squaredNorm()) * Eigen::Matrix<double, 2, 2>::Identity(2, 2);
-  }
-  double gamma(Eigen::Vector2d x) const override { return 0.0; }
-  double eta(Eigen::Vector2d x) const override { return 0.0; }
-  bool EssentialConditionsOnEdge(const lf::mesh::Entity& edge) const override {
-    return edge_sel_(edge);
-  }
-  bool IsImpedanceEdge(const lf::mesh::Entity& /*edge*/) const override {
-    return false;
-  }
-  double f(Eigen::Vector2d /*x*/) const override { return 2.0; }
-  double g(Eigen::Vector2d x) const override {
-    return 0.5 * std::log(x.squaredNorm() + 1);
-  }
-  double h(Eigen::Vector2d x) const override {
-    if (x[0] > x[1]) {
-      return x[0];
-    } else {
-      return x[1];
-    }
-  }
-
- private:
-  /// Selector for Dirichlet boundary conditions
-  lf::refinement::EntityCenterPositionSelector<
-      std::function<bool(Eigen::Vector2d)>>
-      edge_sel_;
-};
-
-TEST(lfe_bvp, bvp_1) {
-  // The solution and its gradient
-  auto u = [](Eigen::Vector2d x) -> double {
-    return (0.5 * std::log(x.squaredNorm() + 1.0));
-  };
-  auto grad_u = [](Eigen::Vector2d x) -> Eigen::Vector2d {
-    return (1.0 / (x.squaredNorm() + 1.0)) * x;
-  };
-  auto bvp_p = std::make_shared<MixedEllipticBVP>();
-  auto errnorms = TestConvergenceEllBVPFESol(4, bvp_p, u, grad_u);
-}
-
 }  // namespace lf::fe::test
